+++ src/2_setup/player_setup.c
@@ -1,34 +1,18 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   settings.h                                         :+:      :+:    :+:   */
+/*   player_setup.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: freddy <freddy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/05/28 08:45:22 by freddy            #+#    #+#             */
-/*   Updated: 2024/05/29 08:33:41 by freddy           ###   ########.fr       */
+/*   Created: 2024/05/29 08:28:05 by freddy            #+#    #+#             */
+/*   Updated: 2024/05/29 08:28:46 by freddy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef SETTINGS_H
-# define SETTINGS_H
+#include "../../include/cub3d.h"
 
-#include <stdbool.h>
-
-// General
-
-# define GREETING true
-# define VERBOSE true
-# define DEBUG true
-
-// Window Setup
-
-# define START_HEIGHT 540
-# define START_WIDTH 960
-# define WINDOW_NAME "cub3d"
-
-// Player Movement Stats
-
-# define MOVEMENT_SPEED 0.5
-
-#endif
+void	setup_player(void)
+{
+	get_player()->transform = get_player()->spawn_transform;
+}