+++ src/3_game_loop/a_game_logic/hooks/loop_hook.c
@@ -6,7 +6,7 @@
 /*   By: freddy <freddy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/28 10:23:28 by freddy            #+#    #+#             */
-/*   Updated: 2024/05/28 10:36:32 by freddy           ###   ########.fr       */
+/*   Updated: 2024/05/29 09:08:54 by freddy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,6 +14,9 @@
 
 void	loop_hook(void *param)
 {
-	write(STDOUT_FILENO, "H", 1);
+	t_entity	*p;
+
+	p = get_player();
+	printf("POS: [%f, %f] - ROT: [%f, %f]: %d\n", p->transform.pos.x, p->transform.pos.y, p->transform.rot.x, p->transform.rot.y, *get_persistent_data()->input_data->map[(int)p->transform.pos.y][(int)p->transform.pos.x]);
 	(void) param;
 }