+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.9.0] - 29.05.2024 - Freddy - `18-player-movement`
+- dysfunctional movement setup, no rotations yet. values dont seem to update. but ive changed so much its committment time
+- Changed seperate vec2s in player entity to use one t_transform instead
+- made scale_transform use a double for precise values & made it take in any vec2 instead of a transform
+- random fixes
+
 ### [0.8.2] - 28.05.2024 - Freddy - `15-validation-flood-fill`
 - made map generation place player pos where generation started, so pretty much the center of the labrinth
 - some map gen fixes & refactorings